tools: roast: refactor as a lib instead of main package
