+++ toolkit/tools/pkg/image/roast/config.go
@@ -1,0 +1,11 @@
+package roast
+
+type Config struct {
+	InputDir       string
+	OutputDir      string
+	ConfigFile     string
+	TmpDir         string
+	ReleaseVersion string
+	Workers        int
+	ImageTag       string
+}
