+++ toolkit/tools/pkg/image/roast/formats/diff.go
