+++ toolkit/tools/pkg/image/roast/formats/initrd.go
